pyproject.toml
[tool.poetry]
name = "quizhub"
version = "0.1.0"
description = "Backend API DRF"
authors = ["sohan"]
license = "NA"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
django-redis = "^5.4.0"
celery = "^5.4.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-leaflet = "^0.30.1"
djangorestframework-api-key = "^3.0.0"
drf-yasg = "^1.21.7"
django = ">=4.2.0,<4.3.0"
setuptools = "^75.1.0"
pillow = "^10.4.0"
djangorestframework-simplejwt = '*'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
